# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Zoptal Platform API
  description: |
    Comprehensive API documentation for the Zoptal microservices platform.
    
    This API provides access to all Zoptal services including authentication,
    project management, AI services, billing, notifications, and analytics.
    
    ## Authentication
    Most endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    API requests are rate limited to prevent abuse. Rate limits are applied per user/IP address.
    
    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.
    
  version: 1.0.0
  contact:
    name: Zoptal API Support
    email: api-support@zoptal.com
    url: https://docs.zoptal.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://zoptal.com/terms
servers:
  - url: https://api.zoptal.com
    description: Production server
  - url: https://staging-api.zoptal.com
    description: Staging server
  - url: http://localhost:4000
    description: Local development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /projects:
    get:
      tags:
        - Projects
      summary: List user projects
      description: Get a list of projects for the authenticated user
      operationId: listProjects
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Filter by project status
          schema:
            type: string
            enum: [active, archived, completed]
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags:
        - Projects
      summary: Create new project
      description: Create a new project for the authenticated user
      operationId: createProject
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
  /projects/{projectId}:
    get:
      tags:
        - Projects
      summary: Get project details
      description: Get detailed information about a specific project
      operationId: getProject
      security:
        - BearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: Project ID
          schema:
            type: string
      responses:
        '200':
          description: Project retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      tags:
        - Projects
      summary: Update project
      description: Update project information
      operationId: updateProject
      security:
        - BearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: Project ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      tags:
        - Projects
      summary: Delete project
      description: Delete a project (soft delete)
      operationId: deleteProject
      security:
        - BearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: Project ID
          schema:
            type: string
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  # AI Services
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        avatar:
          type: string
          format: uri
          description: URL to user's avatar image
        role:
          type: string
          enum: [user, admin, developer]
          default: user
        status:
          type: string
          enum: [active, inactive, suspended]
          default: active
        emailVerified:
          type: boolean
          default: false
        phoneVerified:
          type: boolean
          default: false
        twoFactorEnabled:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, archived, completed]
          default: active
        visibility:
          type: string
          enum: [private, public, team]
          default: private
        owner:
          $ref: '#/components/schemas/UserProfile'
        members:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMember'
        settings:
          $ref: '#/components/schemas/ProjectSettings'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectMember:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
        role:
          type: string
          enum: [owner, admin, member, viewer]
        permissions:
          type: array
          items:
            type: string
        joinedAt:
          type: string
          format: date-time
    ProjectSettings:
      type: object
      properties:
        enableNotifications:
          type: boolean
          default: true
        enableAnalytics:
          type: boolean
          default: true
        defaultAIModel:
          type: string
        apiRateLimit:
          type: integer
          minimum: 100
          maximum: 10000
          default: 1000
    CreateProjectRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        visibility:
          type: string
          enum: [private, public, team]
          default: private
        settings:
          $ref: '#/components/schemas/ProjectSettings'
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [active, archived, completed]
        visibility:
          type: string
          enum: [private, public, team]
        settings:
          $ref: '#/components/schemas/ProjectSettings'
    ProjectListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            projects:
              type: array
              items:
                $ref: '#/components/schemas/Project'
            pagination:
              $ref: '#/components/schemas/PaginationInfo'

    # AI Service Schemas
...
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: string
            field:
              type: string
              description: Field name for validation errors
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: object
          properties:
            code:
              type: string
              default: VALIDATION_ERROR
            message:
              type: string
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  code:
                    type: string
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: BAD_REQUEST
              message: Invalid request parameters
    Unauthorized:
      description: Unauthorized - Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
...
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
...
    ValidationError:
      description: Validation error - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              errors:
                - field: email
                  message: Invalid email format
                  code: INVALID_EMAIL
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred